requirements.txt
streamlit==1.28.0
pandas==2.1.2
altair==5.1.2
pillow==10.1.0
